pci: fix panic if a secondary bus number is changed to out of range of MCFG
Previously config_mmio, the MCFG information, was kept when the bus
number was changed.  The behavior was incorrect because the bus number
might be out of range of the MCFG space.  For example, the MCFG space
is:

MCFG [0] 0000:00-9B (E0000000,9C00000)

and a guest operating system changes a bus number from 0A to C2.  In
this case, previously panic() might be called by the
pci_readwrite_config_mmio() function.

Now config_mmio of a device is temporarily cleared when a bus number
of the device is changed, and set again in the pci_reconnect_device()
function.