svm: add unsafe_nested_virtualization support
If new config vmm.unsafe_nested_virtualization is 1, unsafe nested
virtualization support is enabled.  Only AMD-V is currently supported.

Unsafe means no protection of the host VMM and no hooks of devices.  A
VM on the guest VMM can access everything, including devices and the
host VMM in the RAM, if the guest VMM gives the VM direct access to
the host memory or device I/O.  No para-passthrough drivers work when
a VM on the guest VMM accesses directly.  Safety depends on the guest
VMM.

Previously SVM was concealed in the cpuid instruction emulation.  Now
it is disabled in the MSR emulation (detected as "disabled by BIOS" by
the guest VMM) when vmm.unsafe_nested_virtualization is 0.