thread: use ticketlock in schedule()
The changeset 6d92a514fc59 and d32ef8f9efc7 introduced lock-free
algorithm for scheduling threads.  However a panic "thread:
td_runnable_put failure %u %u" occurred infrequently and the cause of
the panic was unknown.

The lock-free scheduling was introduced because of fairness of
spinlocks and speed.  Before it was introduced, the spinlock of the
thread was sometimes taken by only one CPU running a simple code like
below for more than 10 seconds:

    while (state != expected)
            schedule ();

This changeset reverts the lock-free scheduling to previous spinlock
based implementation, but introduces a ticket lock for fairness, so
the above code still works properly.