thread: add CONFIG_THREAD_1CPU that disables parallel thread processing
Previously a ticketlock used by the schedule() function took a long
time when there were a lot of threads and multiple logical CPUs.  If
the new config option CONFIG_THREAD_1CPU is set, a spinlock is used
instead of the ticket lock, and parallel thread processing is disabled
i.e. threads created by the thread_new() function are executed only
when other logical CPUs are executing none of them.  This option may
improve system performance if there are a lot of threads.