bnx: PCIe capabilities passthrough for virtio-net
Previously if the bnx driver was used with "virtio=1,multifunction=1"
driver options and the device was multifunction, Windows BSoD
occurred.  Apparently it was caued by concealing PCIe device
capabilities of only one instance of a multifunction device.  Making
capabilities passthrough fixes the problem.

MSI and MSI-X capabilities are hidden because MSI or MSI-X support of
the virtio-net device is currently not implemented.