virtio-net: reset ISR status field to zero after read
The ISR status field in the virtio header should show that an
interrupt is generated by this device or not, and should be cleared
after read.  Previously the ISR status field was always read as one.
The previous implementation might confuse a guest operating system
that shared an interrupt with the virtio-net device and other devices,
and might cause interrupt flood.

Now the ISR status field is cleared to zero after read and set after
generating an interrupt.