disconnect firmware drivers before initializing devices in hypervisor drivers
A device has been initialized when BitVisor is loaded if there is a
firmware driver for the device.  A hypervisor driver can conflict with
the firmware driver and cause a system to hang.  Especially hypervisor
drivers that conceal or act as a different device like virtio-net
conflict very often.

This changeset adds an API to disconnect the firmware drivers from the
device, and make hypervisor drivers call it to avoid conflicts.