msr_pass: microcode update passthrough on Intel CPUs like on AMD CPUs
Microcode updates cannot be loaded in VMX non-root operation on Intel
CPUs, even if wrmsr MSR_IA32_BIOS_UPDT_TRIG is set to passthrough.
Previously the wrmsr was just ignored except printing a message, but
Linux >= 4.10 did not boot with microcode updates because it
apparently retries forever while the update revision is not properly
updated.

To make Intel microcode update passthrough, the hypervisor needs to
map or copy guest operating system pages containing the update into
the hypervisor virtual memory before doing the wrmsr, but there is no
way to get the size of the update.  The wrmsr apparently generates a
page fault if pages that CPU wants are not mapped, so the hypervisor
uses page faults to detect which pages are wanted by CPU.