pci: pci_set_bridge_io() ignores hot-plug devices
The pci_set_bridge_io() function sets I/O space address of PCI
bridges.  It is for a virtio device.  The firmware does not have
information about the virtual device, so the I/O space may not be
prepared when the VMM is started.  The function is not compatible to
hot-plug devices, because hot-plug devices are handled in the virtual
world, so the guest operating system (or maybe the firmware) looks at
the virtual device information which needs the I/O space.