acpi_dsdt: apply ObjectList removal in ACPI spec 6.1 Errata A
Previously ObjectList in DefDevice was defined as ObjectList2 to make
If() operator and ZeroOp usable.  Now TermList is used instead of
ObjectList.  If() operator is usable in TermList.

ZeroOp is handled as NullName of NameString in MethodInvocation.  If
there are two or more ZeroOp bytes, the first byte is NullName and
remaining bytes are ZeroOp in TermArgList.