bnx: add hotplugpass option
Currently initialization in the bnx driver is not enough.  The driver
works properly if the firmware initialized the device before starting
the VMM.

Hot plugged device should be initialized properly.  If the bnx driver
is enabled, the driver should initialize the device when the device is
hot plugged because the driver does not support para-passthrough mode.
However the driver currently cannot initialize the device properly.

This changeset introduces "hotplugpass=1" driver option for the bnx
driver which enables passthrough mode for hot plugged devices.  Hot
plugging detection in the driver new() function checks whether the
function is called before or after the "config1" initfunc call.  Hot
plugging detection in the driver reconnect() function reads a ring
buffer address via PCIe configuration space (because the MMIO space
may not be able to be accessed at that time) and compare it with a
value that the driver previously wrote.