nvme: fix problems when queue paring is irregular
Typical queue paring is Submission Queue n is paired to Completion
Queue n, whene n >= 1. However, the specification allows a Submission
Queue to be paired with any Completion Queue. Multiple Submission
Queues can also be associated with a Completion Queue. Current
NVMe implementation does not handle irregular queue paring properly.

There are two problems.
1) Current request_hub assumes Submission Queue n to be paired
   to Completion Queue n by the guest OS. This can results in
   incorrect value writing to a Completion Doorbell Register on
   behalf on the guest.
2) process_comp_queue() does not patch Completion entries properly
   in case of irregular queue mapping causing the incorrect value
   to return to the guest.

To fix the first problem, the number of request_hubs must equal to
the number of Completion Queues. We then add Submission Queues to
a proper request_hub according to the Completion Queue it is
associated with. In this way, we can properly track the number of
waiting requests, and not-yet-acknowledged requests which are
necessary to avoid mixing host requests, and guest requests.

To fix the second problem, The Completion Entry's Queue Head must be
patched with the proper value from its associated Submission Queue
tail value.