exint, vt, svm: clean up exint functions and check blocking by STI on VT-x
Previously VM entry failed panic happened if blocking by STI bit was
set when an external interrupt caused a VM exit.  The bit is
apparently not set on physical machines at an external interrupt VM
exit, but may be set on virtual machines like Linux KVM.  To avoid the
panic, the exint functions are needed to be cleaned up.