panic: clean up code and make it dump all processors if number of processors is more than 6
Previously the panic() function printed registers of each processor of
up to 6 processors.  The reason was the following:

- After the panic() function was called first, it was also called by
  other processors.

- There was a global counter keeping number of calls to the panic()
  function to avoid unexpected panic() calls from itself.

Instead of the current global counter, something like state variable
per processor is needed.  Now IDTR limit is used to track which action
is going in the previous panic() because the IDTR limit is a processor
register and always accessible while the currentcpu structure is not.