mm, pci: remove mapmem_gphys, add mapmem_as and as_dma
The previous mapmem_gphys had the following defects:

- The address passed to the mapmem_gphys was guest physical address of
  the "current" virtual machine.  However the para pass-through driver
  expects that the address is guest physical address of the
  pass-through virtual machine.

- The mapmem_gphys does not call mmio functions while the
  read/write_gphys_b/w/l calls them.

- Because the mapmem_gphys could not be used before initialization of
  the "current" pointer, some para pass-through drivers used the
  mapmem_hphys instead of the mapmem_gphys.

- Para pass-through drivers should be able to map DMA address space to
  support IOMMU.

The new mapmem_as function maps the specified address in the specified
address space.  The address space of the pass-through virtual machine
is defined as "as_passvm".  The para pass-through drivers can use the
"as_passvm" directly to map registers at the PCI base address registers.

DMA address space of every PCI device is defined as "as_dma" in the
pci_device structure.  Currently it is the same as the "as_passvm",
but it can be used for IOMMU support.