pro1000: generate virtio interrupts properly on recent Linux guest
Previously the pro1000 driver allocated an interrupt using the
exint_pass_intr_alloc() function and changed the interrupt number to
the proper number based on MSI-X vectors when an interrupt came.  It
did not work properly because:

- The interrupt was not redirected to a proper processor (on Linux
  4.19.0).

- Priority of interrupts allocated by the exint_pass_intr_alloc()
  function is one.  Task-Priority Class in TPR must be zero to avoid
  blocking the interrupts.  Linux 5.5.8 set TPR to one, so interrupts
  were blocked.

To fix these issues, the pro1000 driver now uses a virtio input
interrupt of the virtio-net device for the real device.