acpi, boot/uefi-loader: modify ACPI tables again when new tables are installed or uninstalled
EFI_ACPI_TABLE_PROTOCOL is a protocol used to install or uninstall
ACPI tables.  Depending on firmware implementation, it might free and
allocate every table.  Therefore the modification in the core/acpi.c
might lost.

To keep the modification,
EFI_BOOT_SERVICES.InstallConfigurationTable() is hooked by a boot
service driver and apply the modification to the new table when it is
called with ACPI_TABLE_GUID or EFI_ACPI_20_TABLE_GUID.
EFI_ACPI_TABLE_PROTOCOL is not hooked because old EFI firmware may not
have the protocol.  The boot service driver is loaded by the
loadvmm.efi containing the binary of the driver when it is needed.

The driver currently provides only one function, acpi_table_mod().
The TableAddr argument can be used for modifying address of a table,
for future use.