nvme: improve 128-byte I/O command size handling for Apple ANS2 Controller
Recent Linux can detect ANS2 controller without a need to set 'ans2_wrapper'.
However, it tells the device that it is going to use 64-byte I/O command size
even though it uses 128-byte I/O command size internally. That means we cannot
rely on the value that the guest provides.

If the guest can detects ANS2 controller without setting 'ans2_wrapper', we
have to check the second command that the guest submits. If the command content
is all zero, we assume that the guest uses 128-byte I/O command size.

Note that Setting 'ans2_wrapper' should force the guest to assume that the
device supports only 64-byte I/O command size. So, it is unnecessary to check.