bnx: ignore firmware's MAC address for Thunderbolt to Ethernet devices
On Mac Mini 2018, we found that the firmware incorrectly assigns the same
MAC address to any NIC devices connected during booting. It seems that
the MAC address from the firmware is intended for the internal NIC device.
In addition, macOS uses the MAC address found the the Thunderbolt to Gigabit
Ethernet device.

To avoid MAC address conflict, we ignore the MAC address assigned by the
firmware for Thunderbolt to Gigabit Ethernet devices.