bnx: fix unexpected delay of received packets
If the bnx driver is used with virtio=1 and
net=pass/ippass/ippassfilter, received packets of the bnx device are
passed to a virtio-net device of the virtual machine.  An interrupt of
the bnx device is handled by the guest operating system as an
interrupt of the virtio-net device.  When an interrupt handler in the
guest operating system reads ISR status, the hypervisor clears the
device interrupt state and handles received packets.  The reason of
clearing first is that if new packets arrive while the hypervisor is
handling the previous packets, another interrupt is expected to handle
the new packets as soon as possible.

However previously the hypervisor sometimes handled received packets
before clearing the device interrupt state or before memory write from
the device.  Doing a dummy read of a device register fixes the issue,
since the read access needs to wait for other transfer completion via
the PCI host bridge.