localapic: BSP ensures all APs are in wait-for-SIPI state after ap_start()
On UEFI environment, APs are started when a guest operating system
sends INIT IPI.  Previously the guest operating system might send SIPI
to a processor which was not in wait-for-SIPI state yet and failed to
start the processor, because BSP continued main thread running while
APs were still initializing.  This patch avoids such situation.