acpi: unifying register initialization/access, starting with reset_reg
gas_write()'s use of access_size is technically not correct. 'access_size'
is only available on ACPI 3.0+. Otherwise, it is undefined. We should not
assume that undefined is 1 byte access. In addition, GAS can also be NULL
on some platforms. We need to verify it before using.

As we plan to use GAS structure for initializing pm1a_cnt and pm_tmr for
MMIO support later, it is a good idea to have unify initialization and
access for ACPI registers.