mm: make alloc_uppermem() use vmm_mem_alloc_uppermem()
The implementation relies on e820 memory information which are initialized
during VMM memory initialization. Let vmm_mem code handle the allocation
logic so that we don't have to expose variables related to upper memory
allocation.

Note that this is an intermediate commit. vmm_mem_*() functions will be
moved to vmm_mem.c later.