efi-app: switch from mingw to clang + lld, prepare for cross-compiling
We can cross-compile our UEFI loader with clang + lld. This is necessary
for multiple platform support later.

As uefi-boot, uefi-boot-login, and dbgsh-uefi are similar, we now have
Makefile.uefi.common for sharing common parameters.

Note that we produce object files, and link them later instead of creating
efi file directly. This is to make sure that the .d files are generated
correctly for file dependency check.

In addition, we need the following for new clang compatibility:

-nostdlib (LDFLAGS)
    We need tell the linker not to link standard libs explicitly.
-fno-PIE (CLFAGS)
    The compiler complains that our target does not support PIE.
-Wno-microsoft-static-assert (CFLAGS)
    This is to suppress warning during compiling about not including
    assert.h from EDK2 headers.