core: use builtin atomic_cmpxchg*() in x86 specific code
asm_lock_cmpxchg*() always spills 'setne' instruction. However, in most
cases, we don't need to check the result of the operation. This can be a
little inefficient.

Replacing asm_lock_cmpxchg*() with compiler builtin atomic_cmpxchg*()
allows the compiler to optimize in case we don't need to check the result
of the operation. This should lead to a more efficient code.

Note that the return value of asm_lock_cmpxchg*() and atomic_cmpxchg*()
are inverse. As a result, some part of the code need some adjustment
regarding the return value.