vt, cpuid: conceal tpause, umonitor and umwait if not usable in the guest
WAITPKG bit of CPUID shows whether tpause, umonitor and umwait are
supported.  Setting "enable user wait and pause" in Secondary
Processor-Based VM-Execution Controls allows the guest to use the
instructions.

Previously, when the WAITPKG bit showed the instructions were
supported, setting the "enable user wait and pause" was expected to be
allowed.  However this was not true on virtualized environments.  A
guest operating system running on BitVisor on such environments might
catch an unexpected exception when it used one of the instructions.
Concealing the WAITPKG bit avoids the issue.