mm: add exec argument to mm_process_map_shared_physpage()
Executable bit is necessary for shared system entry page and program's
code. Future architecture implementation is going to rely on this.

Note that we don't change x86 implementation (always executable) in this
patch. We want to leave current x86 implementation as untouched as we can.
Proper executable bit handling can be done later.