drivers/usb/xhci: fix "ERDP out of bounds" error when running on QEMU
When running BitVisor on QEMU, QEMU performs an ERDP range check
unlike real devices, and outputs the error "ERDP out of bounds" while
the guest is starting up.

The cause was patch_er_dq_ptr(). This function returns the host's ERDP
address if "host->run" is true, but returns the guest's ERDP address
if "host->run" is false. take_control_erst() called this function
internally and expected the host's address. However,
take_control_erst() was actually called when "host->run" was false.

To fix this bug, this patch changes the order in which
take_control_erst() is called so that the internal patch_er_dq_ptr()
returns the host's ERDP address.