Makefile: make subfolder objects comes before current folder objects
This should affect only process building.

Processes in BitVisor are finally linked with the process/lib/'s static
library. Linker links the first symbol it founds in the static library. We
later want to prioritize architecture-dependent functions over the
architecture-independent ones for a given symbol. As a result, we have to
change the order in $(objs) by making objects in subfolders come before
the current folder during creating the static library.