uhci: fix incorrect use of builtin_ffs()
The original implementation of __ffs() used 'bsf' x86 instruction. The
value the previous implementation returned was 0-based index if '1' bit
exists or undefined otherwise. However, The compiler builtin ffs function
returns 1-based index if '1' bit exists or 0 otherwise. As a result,
off-by-1 error occurs, and results in crash.

To preserve the original behavior of __ffs(), we have to subtract the
value from builtin_ffs() by 1. There is no need to check for return value
of builtin_ffs() as the previous __ffs() expected the caller to check the
input value before using the function.

To avoid confusion in the future, it is good to add a comment on what to
expect from builtin_ffs().