xhci: process the current state of the event ring properly
According to the xHCI spec, the controller resets its internal PCS
flag and base address of the event ring when a software writes to
ERSTBA. The software also should set its internal states properly.
Previously, the driver did not properly reinitialize:
* segment index
* event ring index
* CSS flag
when the guest wrote to ERSTBA. In addition, the driver set the wrong
state when the driver created VMM's event ring.

This commit fixes the above mentioned problems. In addition, the
driver resets the guest's event ring related data when the guest
writes to ERSTBA.