xhci: reallocate the VMM's event ring if the size of the ring changes
Previously the driver did not recheck the size of the event ring when
the controller stopped and restarted.

For example, the UEFI xHCI driver initializes the controller to use
512 entries event ring. After that, the guest OS like Linux is booted,
and its xHCI driver initializes the controller to use 256 entries
event ring, the BitVisor's xHCI driver breaks.

This commit ensures that the driver checks for the guest event ring
size and reallocates accordingly.