xhci: remove "state_saved" member and panic when CRS bit is set
The "state_saved" member in "xhci_host" struct was rarely used.
Handling the CRS bit set event was also not actually implemented.

This commit removes the "state_saved" member and triggers a panic when
CRS bit is set. With this commit, we no longer need to take the
"state_saved" into account. It is also clearer that the driver
currently does not handle the CRS bit set event.