xhci: rename "xhci_hc_reset()" and call it at the proper time
"xhci_hc_reset()" was a function that freed the VMM's data and
unmapped and freed the guest's data. However, it didn't reset the
controller. This name led to misunderstandings. Additionally, this
function was only called when the VMM detected a controller error.
This implied that these release routines were not called during
normal operation.

This commit renames "xhci_hc_reset()" to "xhci_release_data()" so that
the function name reflects what it does. In addition,
"xhci_release_data()" is called just before the controller starts to
clean up previous internal states/data.