pci: PCI PMIO is only relevant to x86
On non-x86 machine, it is possible that each PCI segment/domain has its
own I/O address space as MMIO memory. Checking whether we can use PMIO
based on core_io is not relevant.

PMIO is only relevant to x86 machine. Therefore, it is simpler to not
support PMIO on other architectures.