aarch64: initial implementation
This is the initial implementation for AArch64 BitVisor. The current stage
of the implementation is that it can boot the Linux kernel with SMP and a
functional network through virtio_net on QEMU. It currently boots only on
the UEFI environment with ACPI. It requires the platform to support PSCI
for secondary CPU core boot-up.

The implementation targets Armv8.1. We require Virtualization Host
Extension (VHE) to simplify MMU implementation, and make process
implementation similar to the original x86 version.

The AArch64 implementation is documented in docs/aarch64/overview.md. It
provides an overview of how AArch64 BitVisor works from initialization to
returning the control back to the guest.

See docs/getting_started.md for instructions to run AArch64 BitVisor on
QEMU.