core: remove vpn_ve code
vpn_ve requires a driver on the guest OS. The driver is not available
publicly. vpn_ve was introduced to test VPN module in BitVisor without any
para-passthrough network drivers. However, there are many para-passthrough
network drivers that work with VPN module now. So, it is ok to remove
vpn_ve code.