acpi_dsdt: simplify and improve parsemain() make it more deterministic
At introducing namespace search, in parsemain(), 'r' is cleared
unconditionally after calling getbuf(d). The previous implementation
expected 'r', once set, to remain as it was and returned when getbuf(d)
returns NULL in 'r != NULL' path to clean up parsing data. The
unconditionally clearing of 'r' can cause the code go to the error case in
'r == NULL' path on some machines.

This patch simplifies parsemain() and make it more deterministic. The
following is the overview of changes.

In case of no search string, once d->c == d->end, we can record the
current d->cur, free remaining parsing data, and return the recorded
d->cur without relying on goto loop.

In case of search string, if d->c == d->end, it means we reach the end of
the search. We need to free the recorded d->cur too before returning
NULL as the string the caller looks for is not found. Otherwise, if the
search string is found, we record the current d->cur, free remaining
parsing data, and return the recorded d-cur.