aarch64/gic: making reading of IAR visible immediately
We read Interrupt Acknowledge Register(IAR) to get the current pending
interrupt ID (INTID). It is a good idea to add isb following by dsb to
make the effect of reading visible immediately. This is to follow what the
specification says.

According to the GICv3 specification:
"The effects of reading ICC_IAR0_EL1 and ICC_IAR1_EL1 on the state of a
returned INTID are not guaranteed to be visible until after the execution
of a DSB."

"An ISB or other context synchronization operation must precede the DSB to
ensure visibility of System register writes."