aarch64/mmio: avoid using memcpy() for accessing mapped guest address
An address passed to mmio_gphys_access() can be a MMIO address. It is
better to access the address with access size of 1, 2, 4, or 8 bytes
explicitly. It is because it is possible for some devices to have strict
access size requirement. For example, a device can expect 4-byte access to
a register. Reading the register 1-byte at a time can result in wrong
value (all 0 rather than the actual value of the register for example).