core/mm: add expect_writable parameter to mm_process_arch_virt_to_phys()
When a process or VMM sends a buffer to others, it is the senders who
specify the buffer write permission. However, it is also possible that
senders make a mistake by specifying write permission to a read-only
buffer. We can avoid this problem, and make it easier to debug by adding
expect_writable parameter to mm_process_arch_virt_to_phys(). The actual
implementation of virtual to physical address translation takes the
parameter into account, and returns an error in case of expectation
mismatch.