aarch64/thread: forget save and restore some necessary system registers
The current implementation is not enough to handle voluntary wait from
processes. To handle this, the thread scheduling needs to save and restore
the HCR_EL2, SPSR_EL2, ELR_EL2, and TPIDR_EL0 as well. When an exception
entry is from a BitVisor process:
	- HCR_TGE is set (necessary for launching a process from EL2 with
	  Virtualization Host Extensions (VHE) enabled)
	- SPSR_EL2 contains PSTATE of the process
	- ELR_EL2 contains the address to return to the process
	- TPIDR_EL0 can be modified as it can be used by processes
When a thread is scheduled back to resume its process execution, it can
return to the process with the correct system register context.