usb: usb_hook_process() supports usb_hook_register() in hook callback properly
Previously, when usb_hook_register() was called during a hook
callback, the registered hook might or might not be called during the same
loop of usb_hook_register().  The reason was that a next pointer,
which was in a local variable of usb_hook_process() during the hook
callback, might be modified by usb_hook_register().  Therefore, the
behavior might be different depending on the position of the internal
hook list.