ehci: clean up register handler and handle doorbell properly
The 'doorbell' member of 'struct ehci_host' controls the Interrupt on
Async Advance Doorbell bit in the USB Command Register, which may not
be set by a guest driver.  Previously, if the guest driver did not set
it, the VMM would internally set the bit in a 'buf32' variable, but
the value was unused.

The previous register handler lacked clarity on how the 'buf32'
variable was used.  This changeset refactors the handler and ensures
the updated 'buf32' value is applied properly.